move mark_disabled inside ed25519_verify for gpu case
