+++ core/src/sigverify_stage.rs
@@ -45,8 +45,7 @@ pub struct DisabledSigVerifier {}
 
 impl SigVerifier for DisabledSigVerifier {
     fn verify_batch(&self, mut batch: Vec<Packets>) -> Vec<Packets> {
-        let r = sigverify::ed25519_verify_disabled(&batch);
-        sigverify::mark_disabled(&mut batch, &r);
+        sigverify::ed25519_verify_disabled(&mut batch);
         batch
     }
 }
